refactor: model revenue with BigDecimal

+++ src/main/java/com/ecommerce/service/OrderAnalyticsService.java
@@ -1,9 +1,10 @@
 package com.ecommerce.service;
 
 import com.ecommerce.model.Order;
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface OrderAnalyticsService {
-    double calculateTotalRevenue(List<Order> orders);
+    BigDecimal calculateTotalRevenue(List<Order> orders);
     int countUniqueProducts(List<Order> orders);
 }
